산책로 삭제 수정 및 api 형식(ResponseDto) 통일 작업

+++ src/main/java/com/promenade/promenadeapp/controller/User/UserRoadController.java
@@ -3,7 +3,6 @@ package com.promenade.promenadeapp.controller.User;
 import com.promenade.promenadeapp.domain.User.User;
 import com.promenade.promenadeapp.domain.User.UserRoad;
 import com.promenade.promenadeapp.dto.ResponseDto;
-import com.promenade.promenadeapp.dto.RoadPathResponseDto;
 import com.promenade.promenadeapp.dto.UserRoadRequestDto;
 import com.promenade.promenadeapp.dto.UserRoadResponseDto;
 import com.promenade.promenadeapp.service.User.UserRoadService;
@@ -30,7 +29,7 @@ public class UserRoadController {
         List<UserRoad> userRoads = userRoadService.getUserRoads(googleId);
         if (userRoads.isEmpty()) {
             ResponseDto response = ResponseDto.builder()
-                    .error("사용자의 커스텀 산책로가 없습니다.")
+                    .error("사용자의 커스텀 산책로가 없습니다. userGoogleId = " + googleId)
                     .build();
             return ResponseEntity.badRequest().body(response);
         }
@@ -72,23 +71,18 @@ public class UserRoadController {
         }
     }
 
-    @DeleteMapping
+    @DeleteMapping("/id/{id}")
     public ResponseEntity<?> deleteUserRoad(@AuthenticationPrincipal String googleId,
-                                            @RequestBody UserRoadRequestDto requestDto) {
+                                            @PathVariable Long id) {
         try {
-            User foundUserByGoogleId = userService.findByGoogleId(googleId);
-            Long foundRoadId = userRoadService.findByTrailName(requestDto.getTrailName()).getId();
-
-            UserRoad userRoad = UserRoad.builder()
-                    .id(foundRoadId) // trailName(unique key)으로 찾은 road id
-                    .userGoogleId(googleId) // googleId 추가
-                    .trailName(requestDto.getTrailName())
-                    .description(requestDto.getDescription())
-                    .distance(requestDto.getDistance())
-                    .startAddr(requestDto.getStartAddr())
-                    .trailPoint(requestDto.getTrailPoint())
-                    .user(foundUserByGoogleId) // user 추가
-                    .build();
+            Long userId = userService.getUserIdByGoogleId(googleId);
+            UserRoad userRoad = userRoadService.findById(id);
+            if (userId != userRoad.getUser().getId()) {
+                ResponseDto response = ResponseDto.builder()
+                        .error("요청한 산책로 id가 로그인한 당신의 산책로가 아닙니다. roadId = " + id)
+                        .build();
+                return ResponseEntity.badRequest().body(response);
+            }
 
             List<UserRoad> userRoads = userRoadService.deleteUserRoad(userRoad);
